Add Linux script to limit outgoing bandwidth used by Reddcoin
This commit adds a Linux script that uses tc and iptables to
limit outgoing bandwidth to Reddcoin nodes. All tcp connections with
a source or destination port of 45444 are limited to the chosen rate.
It does not alter the incoming bandwidth. Additionally, outgoing
bandwidth to a host on a local LAN (defined by the LOCALNET variable)
is not limited.